♻️ Extract Fallback installer to installer.fallback.ts
