+++ src/installer.fallback.ts
@@ -1,0 +1,130 @@
+import {
+  promptUserOnConflict,
+  promptUserToInstallOrCancelOnReachingFallback,
+} from "./dialogs";
+import { filesUnder, FileTree, FILETREE_ROOT, Glob, sourcePaths } from "./filetree";
+import {
+  FallbackLayout,
+  InvalidLayout,
+  LayoutToInstructions,
+  MaybeInstructions,
+  NoInstructions,
+} from "./installers.layouts";
+import { instructionsForSameSourceAndDestPaths } from "./installers.shared";
+import { InstallerType, InstallDecision } from "./installers.types";
+import { exhaustiveMatchFailure } from "./installers.utils";
+import {
+  VortexApi,
+  VortexInstallResult,
+  VortexLogFunc,
+  VortexTestResult,
+  VortexWrappedInstallFunc,
+  VortexWrappedTestSupportedFunc,
+} from "./vortex-wrapper";
+
+export const findFallbackFiles = (fileTree: FileTree): string[] =>
+  filesUnder(FILETREE_ROOT, Glob.Any, fileTree);
+
+export const detectFallbackLayout = (_fileTree: FileTree): boolean => true;
+
+export const fallbackLayout: LayoutToInstructions = (
+  _api: VortexApi,
+  _modName: string,
+  fileTree: FileTree,
+): MaybeInstructions => {
+  if (!detectFallbackLayout(fileTree)) {
+    throw new Error("Should never get here");
+  }
+
+  const allTheFiles = findFallbackFiles(fileTree);
+
+  return {
+    kind: FallbackLayout.Unvalidated,
+    instructions: instructionsForSameSourceAndDestPaths(allTheFiles),
+  };
+};
+
+export const useFallbackOrFail = (
+  api: VortexApi,
+  installerType: InstallerType,
+  fileTree: FileTree,
+  installDecision: InstallDecision,
+): Promise<VortexInstallResult> => {
+  switch (installDecision) {
+    case InstallDecision.UserWantsToCancel: {
+      const message = `${installerType}: user chose to cancel installation on conflict`;
+      api.log(`info`, message);
+      api.log(`debug`, `Input files: `, sourcePaths(fileTree));
+      return Promise.reject(new Error(message));
+    }
+    case InstallDecision.UserWantsToProceed: {
+      api.log(`info`, `${installerType}: user chose to continue installation`);
+      api.log(`info`, `${installerType}: using fallback layout to install everything`);
+
+      const fallbackInstructions = fallbackLayout(api, undefined, fileTree);
+
+      if (
+        fallbackInstructions === InvalidLayout.Conflict ||
+        fallbackInstructions === NoInstructions.NoMatch
+      ) {
+        return Promise.reject(
+          new Error(
+            `Fallback layout failed, should never get here: ${fallbackInstructions}`,
+          ),
+        );
+      }
+
+      api.log(`info`, `${installerType}: instructions generated by fallback installer`);
+      api.log(`debug`, `Instructions`, fallbackInstructions.instructions);
+
+      return Promise.resolve({
+        instructions: fallbackInstructions.instructions,
+      });
+    }
+    default: {
+      return exhaustiveMatchFailure(installDecision);
+    }
+  }
+};
+
+export const useFallbackOrFailBasedOnUserDecision = async (
+  api: VortexApi,
+  installerType: InstallerType,
+  fileTree: FileTree,
+): Promise<VortexInstallResult> => {
+  const installDecision = await promptUserOnConflict(
+    api,
+    installerType,
+    sourcePaths(fileTree),
+  );
+
+  return useFallbackOrFail(api, installerType, fileTree, installDecision);
+};
+
+export const testForFallback: VortexWrappedTestSupportedFunc = (
+  _api: VortexApi,
+  log: VortexLogFunc,
+  files: string[],
+  _fileTree: FileTree,
+): Promise<VortexTestResult> => {
+  log("debug", "Fallback installer received Files: ", files);
+  return Promise.resolve({
+    supported: true,
+    requiredFiles: [],
+  });
+};
+
+export const installFallback: VortexWrappedInstallFunc = async (
+  api: VortexApi,
+  _log: VortexLogFunc,
+  _files: string[],
+  fileTree: FileTree,
+  _destinationPath: string,
+): Promise<VortexInstallResult> => {
+  const installDecision = await promptUserToInstallOrCancelOnReachingFallback(
+    api,
+    sourcePaths(fileTree),
+  );
+
+  return useFallbackOrFail(api, InstallerType.Fallback, fileTree, installDecision);
+};
