+++ src/installers.ts
@@ -65,7 +65,6 @@ import {
   InvalidLayout,
   CetLayout,
   RedscriptLayout,
-  FallbackLayout,
   AsiLayout,
   KNOWN_TOPLEVEL_DIRS,
 } from "./installers.layouts";
@@ -79,12 +78,7 @@ import {
   useAllMatchingLayouts,
   makeSyntheticName,
 } from "./installers.shared";
-import {
-  showArchiveInstallWarning,
-  showRed4ExtReservedDllErrorDialog,
-  promptUserOnConflict,
-  promptUserToInstallOrCancelOnReachingFallback,
-} from "./dialogs";
+import { showArchiveInstallWarning, showRed4ExtReservedDllErrorDialog } from "./dialogs";
 import {
   testForCetCore,
   installCetCore,
@@ -97,15 +91,15 @@ import {
   testCoreWolvenKitCli,
   installCoreWolvenkit,
 } from "./installer.core";
-import { exhaustiveMatchFailure, trueish } from "./installers.utils";
+import { trueish } from "./installers.utils";
 import { GAME_ID } from "./index.metadata";
-import {
-  InstallDecision,
-  Installer,
-  InstallerType,
-  InstallerWithPriority,
-} from "./installers.types";
+import { Installer, InstallerType, InstallerWithPriority } from "./installers.types";
 import { installCoreTweakXL, TestForCoreTweakXL } from "./installer.core-tweak-xl";
+import {
+  useFallbackOrFailBasedOnUserDecision,
+  testForFallback,
+  installFallback,
+} from "./installer.fallback";
 
 // Ensure we're using win32 conventions
 const path = win32;
@@ -183,115 +177,6 @@ const allCanonicalArchiveOnlyFiles = (files: string[]) =>
 // other parts from the simpler installers.
 //
 //
-
-// Fallback
-
-const findFallbackFiles = (fileTree: FileTree): string[] =>
-  filesUnder(FILETREE_ROOT, Glob.Any, fileTree);
-
-const detectFallbackLayout = (_fileTree: FileTree): boolean => true;
-
-const fallbackLayout: LayoutToInstructions = (
-  _api: VortexApi,
-  _modName: string,
-  fileTree: FileTree,
-): MaybeInstructions => {
-  if (!detectFallbackLayout(fileTree)) {
-    throw new Error("Should never get here");
-  }
-
-  const allTheFiles = findFallbackFiles(fileTree);
-
-  return {
-    kind: FallbackLayout.Unvalidated,
-    instructions: instructionsForSameSourceAndDestPaths(allTheFiles),
-  };
-};
-
-const useFallbackOrFail = (
-  api: VortexApi,
-  installerType: InstallerType,
-  fileTree: FileTree,
-  installDecision: InstallDecision,
-): Promise<VortexInstallResult> => {
-  switch (installDecision) {
-    case InstallDecision.UserWantsToCancel: {
-      const message = `${installerType}: user chose to cancel installation on conflict`;
-      api.log(`info`, message);
-      api.log(`debug`, `Input files: `, sourcePaths(fileTree));
-      return Promise.reject(new Error(message));
-    }
-    case InstallDecision.UserWantsToProceed: {
-      api.log(`info`, `${installerType}: user chose to continue installation`);
-      api.log(`info`, `${installerType}: using fallback layout to install everything`);
-
-      const fallbackInstructions = fallbackLayout(api, undefined, fileTree);
-
-      if (
-        fallbackInstructions === InvalidLayout.Conflict ||
-        fallbackInstructions === NoInstructions.NoMatch
-      ) {
-        return Promise.reject(
-          new Error(
-            `Fallback layout failed, should never get here: ${fallbackInstructions}`,
-          ),
-        );
-      }
-
-      api.log(`info`, `${installerType}: instructions generated by fallback installer`);
-      api.log(`debug`, `Instructions`, fallbackInstructions.instructions);
-
-      return Promise.resolve({
-        instructions: fallbackInstructions.instructions,
-      });
-    }
-    default: {
-      return exhaustiveMatchFailure(installDecision);
-    }
-  }
-};
-
-const useFallbackOrFailBasedOnUserDecision = async (
-  api: VortexApi,
-  installerType: InstallerType,
-  fileTree: FileTree,
-): Promise<VortexInstallResult> => {
-  const installDecision = await promptUserOnConflict(
-    api,
-    installerType,
-    sourcePaths(fileTree),
-  );
-
-  return useFallbackOrFail(api, installerType, fileTree, installDecision);
-};
-
-export const testAnyOtherModFallback: VortexWrappedTestSupportedFunc = (
-  _api: VortexApi,
-  log: VortexLogFunc,
-  files: string[],
-  _fileTree: FileTree,
-): Promise<VortexTestResult> => {
-  log("debug", "Fallback installer received Files: ", files);
-  return Promise.resolve({
-    supported: true,
-    requiredFiles: [],
-  });
-};
-
-export const installAnyModFallback: VortexWrappedInstallFunc = async (
-  api: VortexApi,
-  _log: VortexLogFunc,
-  _files: string[],
-  fileTree: FileTree,
-  _destinationPath: string,
-): Promise<VortexInstallResult> => {
-  const installDecision = await promptUserToInstallOrCancelOnReachingFallback(
-    api,
-    sourcePaths(fileTree),
-  );
-
-  return useFallbackOrFail(api, InstallerType.Fallback, fileTree, installDecision);
-};
 
 //
 //
@@ -1652,8 +1537,8 @@ const installers: Installer[] = [
   {
     type: InstallerType.Fallback,
     id: InstallerType.Fallback,
-    testSupported: testAnyOtherModFallback,
-    install: installAnyModFallback,
+    testSupported: testForFallback,
+    install: installFallback,
   },
 ];
